Revert "[feat] 스케줄 작업 DB 저장 및 프로그램 시작 시 실행 (#70)"
This reverts commit f605e61ffdac8fd0061e9c9290379e432d0b31d6.
