+++ src/main/java/com/fifo/ticketing/domain/book/entity/Book.java
@@ -51,10 +51,6 @@ public class Book extends BaseDateEntity {
     @Column(nullable = false)
     private Integer quantity;
 
-    @OneToOne
-    @JoinColumn(name = "task_id")
-    private BookScheduledTask scheduledTask;
-
     public void canceled() {
         this.bookStatus = bookStatus.CANCELED;
     }
