--- src/main/java/com/fifo/ticketing/domain/book/entity/TaskStatus.java
@@ -1,5 +1,0 @@
-package com.fifo.ticketing.domain.book.entity;
-
-public enum TaskStatus {
-    PENDING, COMPLETED, CANCELED
-}
