+++ src/main/java/com/fifo/ticketing/domain/book/service/BookService.java
@@ -8,7 +8,6 @@ import com.fifo.ticketing.domain.book.mapper.BookMapper;
 import com.fifo.ticketing.domain.book.entity.Book;
 import com.fifo.ticketing.domain.book.entity.BookSeat;
 import com.fifo.ticketing.domain.book.repository.BookRepository;
-import com.fifo.ticketing.domain.book.repository.BookScheduleRepository;
 import com.fifo.ticketing.domain.book.repository.BookSeatRepository;
 import com.fifo.ticketing.domain.performance.entity.Performance;
 import com.fifo.ticketing.domain.performance.repository.PerformanceRepository;
@@ -21,8 +20,11 @@ import com.fifo.ticketing.global.exception.AlertDetailException;
 import com.fifo.ticketing.global.exception.ErrorCode;
 import com.fifo.ticketing.global.exception.ErrorException;
 import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.util.Date;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.scheduling.TaskScheduler;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -31,7 +33,6 @@ import java.util.List;
 import static com.fifo.ticketing.global.exception.ErrorCode.NOT_FOUND_MEMBER;
 import static com.fifo.ticketing.global.exception.ErrorCode.NOT_FOUND_PERFORMANCE;
 
-@Slf4j
 @Service
 @RequiredArgsConstructor
 public class BookService {
@@ -41,7 +42,10 @@ public class BookService {
     private final PerformanceRepository performanceRepository;
     private final SeatRepository seatRepository;
     private final BookSeatRepository bookSeatRepository;
-    private final BookScheduleManager bookScheduleManager;
+
+    @Qualifier("taskScheduler")
+    private final TaskScheduler taskScheduler;
+    private final BookCancelService bookCancelService;
 
     @Transactional
     public Long createBook(Long performanceId, Long userId, BookCreateRequest request) {
@@ -76,9 +80,10 @@ public class BookService {
 
         Long bookId = book.getId();
 
-        LocalDateTime runTime = LocalDateTime.now().plusMinutes(5);
+        LocalDateTime runTime = LocalDateTime.now().plusMinutes(1);
+        Date triggerTime = Date.from(runTime.atZone(ZoneId.systemDefault()).toInstant());
 
-        bookScheduleManager.scheduleCancelTask(bookId, runTime);
+        taskScheduler.schedule(() -> bookCancelService.cancelIfUnpaid(bookId), triggerTime);
 
         return bookId;
     }
@@ -124,12 +129,6 @@ public class BookService {
         return bookId;
     }
 
-
-    @Transactional
-    public List<Book> cancelAllBook(Performance performance) {
-        bookRepository.cancelAllByPerformance(performance, BookStatus.ADMIN_REFUNDED, BookStatus.PAYED);
-        return bookRepository.findAllByPerformanceAndBookStatus(performance, BookStatus.ADMIN_REFUNDED);
-    }
 
     @Transactional
     public List<BookedView> getBookedList(Long userId) {
