+++ src/main/java/com/fifo/ticketing/global/exception/ErrorCode.java
@@ -29,7 +29,6 @@ public enum ErrorCode {
     NOT_FOUND_AUTH("AUTH-002", "잘못된 인증번호 입니다.", NOT_FOUND),
     NOT_FOUND_PROVIDER("PROVIDER-001", "지원하지 않는 플랫폼 서비스입니다.", NOT_FOUND),
     NOT_FOUND_BOOK("BOOK-001", "예약 정보가 존재하지 않습니다.", NOT_FOUND),
-    NOT_FOUND_SCHEDULE("SCHEDULE-001", "존재하지 않는 스케줄 업무입니다.", NOT_FOUND),
     UNAUTHORIZED_REQUEST("AUTH-003", "올바른 요청이 아닙니다. 로그인을 해주세요.", UNAUTHORIZED),
     INVALID_DATETIME_PERIOD("DATETIME-001", "공연 종료 시간은 공연 시작 시간 이후여야 합니다.", BAD_REQUEST),
     INVALID_DATETIME_TYPE("DATETIME-002", "날짜 형식이 올바르지 않습니다.", BAD_REQUEST);
