--- src/main/java/com/fifo/ticketing/global/util/DateTimeUtil.java
@@ -1,13 +1,0 @@
-package com.fifo.ticketing.global.util;
-
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.util.Date;
-
-public class DateTimeUtil {
-
-    public static Date toDate(LocalDateTime localDateTime) {
-        return Date.from(localDateTime.atZone(ZoneId.systemDefault()).toInstant());
-    }
-
-}
